fix for errbit error undefined method is_kit for nil class in order_item.rb

+++ app/models/order_item.rb
@@ -207,9 +207,11 @@ class OrderItem < ActiveRecord::Base
 
   def remove_order_item_kit_products
     result = true
-    if self.product.is_kit == 1
-      self.order_item_kit_products.each do |kit_product|
-        kit_product.destroy
+    unless self.product.nil?
+      if self.product.is_kit == 1
+        self.order_item_kit_products.each do |kit_product|
+          kit_product.destroy
+        end
       end
     end
     result
